logging and decoding

+++ lib/i2c_utils/i2c_utils.h
@@ -11,7 +11,7 @@ class I2CUtils {
         void init(uint8_t device_addr, int i2c_speed);
         uint8_t read_register(uint8_t reg_addr);
         void write_register(uint8_t reg_addr, uint8_t value);
-        void read_registers(uint8_t reg_addr, uint8_t* buffer, int num_bytes);
+        void read_registers(uint8_t reg_addr, uint8_t* buffer, size_t num_bytes);
 };
